backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP client for fetching CSVs
httpx>=0.26.0

# OpenAI SDK (works with Azure OpenAI)
openai>=1.10.0

# Pydantic for data validation
pydantic>=2.5.0

# Environment variable management
python-dotenv>=1.0.0
